requirements.txt
streamlit
pymupdf
python-docx
pandas
re
io